+++ api/src/main/java/com/github/retrooper/packetevents/PacketEventsAPI.java
@@ -18,9 +18,8 @@
 
 package com.github.retrooper.packetevents;
 
-import com.github.retrooper.packetevents.event.BaseEventManager;
 import com.github.retrooper.packetevents.event.EventManager;
-import com.github.retrooper.packetevents.event.RootEventManager;
+import com.github.retrooper.packetevents.event.inheritable.GlobalEventManager;
 import com.github.retrooper.packetevents.injector.ChannelInjector;
 import com.github.retrooper.packetevents.manager.player.PlayerManager;
 import com.github.retrooper.packetevents.manager.protocol.ProtocolManager;
@@ -36,7 +35,7 @@ import com.github.retrooper.packetevents.util.updatechecker.UpdateChecker;
 import java.util.logging.Logger;
 
 public abstract class PacketEventsAPI<T> {
-    private final EventManager eventManager = new RootEventManager();
+    private final EventManager eventManager = new GlobalEventManager();
     private final PacketEventsSettings settings = new PacketEventsSettings();
     private final UpdateChecker updateChecker = new UpdateChecker();
     private final LogManager logManager = new LogManager();
